backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow==11.0.0
numpy==1.26.4
# scikit-learn==1.5.2
scikit-learn==1.6.1
joblib==1.4.2
# tensorflow-cpu==2.15.1
tensorflow-cpu==2.19.0